fix makefile
