+++ .gitignore
@@ -14,3 +14,4 @@ coverage.txt
 
 bin/*
 config.yaml
+api/statik/
