+++ Makefile
@@ -1,13 +1,21 @@
-INTEGRATED_STATIC_FILE?=1
+EMBED_FRONT?=1
 BUILD_TAGS?=
 
-ifeq ($(INTEGRATED_STATIC_FILE), 1)
-	BUILD_TAGS:=integrated $(BUILD_TAGS)
+ifeq ($(EMBED_FRONT), 1)
+	BUILD_TAGS:=embed_front $(BUILD_TAGS)
 endif
 
 .PHONY: $(notdir $(abspath $(wildcard cmd/*/)))
-$(notdir $(abspath $(wildcard cmd/*/))): build-web generate
+$(notdir $(abspath $(wildcard cmd/*/))): before build-web generate
 	go build -tags "$(BUILD_TAGS)" -o bin/$$(go env GOARCH)-$$(go env GOOS)/$@ ./cmd/$@
+ifneq ($(EMBED_FRONT), 1)
+	cp -R web/build/ bin/$$(go env GOARCH)-$$(go env GOOS)/dist
+endif
+
+.PHONY: before
+before:
+	GO111MODULE=off go get github.com/rakyll/statik
+	cd ./web && yarn install && cd ../
 
 .PHONY: clean
 clean:
@@ -16,7 +24,7 @@ clean:
 
 .PHONY:
 build-web:
-	cd ./web && yarn install && yarn build && cd ../
+	cd ./web && yarn build && cd ../
 
 .PHONY: generate
 generate:
