--- api/static_external.go
@@ -1,14 +1,0 @@
-// +build !integrated
-
-package api
-
-import (
-	"net/http"
-
-	"github.com/samaritan-proxy/sash/logger"
-)
-
-func staticFileHandler() http.Handler {
-	logger.Infof("use external mode")
-	return http.FileServer(http.Dir("."))
-}
